Closes #470
