+++ examples/file_operation/CMakeLists.txt
@@ -1,0 +1,14 @@
+add_definitions(-DPENGUINV_ENABLED_PNG_SUPPORT)
+add_executable(example_file_operation
+    ${CMAKE_CURRENT_LIST_DIR}/example_file_operation.cpp
+    ${LIB_DIR}/image_function_helper.cpp
+    ${LIB_DIR}/image_function.cpp
+    ${LIB_DIR}/file/file_image.cpp
+    ${LIB_DIR}/file/bmp_image.cpp
+    ${LIB_DIR}/file/png_image.cpp
+    ${LIB_DIR}/file/jpeg_image.cpp)
+set_target_properties(example_file_operation PROPERTIES
+  ENABLE_EXPORTS 1
+)
+target_link_libraries(example_file_operation
+    PRIVATE example_features_options example_features_warnings png jpeg)
