+++ examples/file_operation/example_file_operation.cpp
@@ -1,0 +1,63 @@
+// Example application of library utilization for bitmaps
+#include <iostream>
+#include "../../src/image_buffer.h"
+#include "../../src/image_function.h"
+#include "../../src/file/file_image.h"
+
+int main( int argc, char * argv[] )
+{
+    std::vector<std::string> filePaths;
+
+    if ( argc > 1 ) // Check input data
+    {
+        for ( int i = 1; i < argc; ++i)
+            filePaths.push_back(argv[i]);
+    }
+
+    else
+    {
+        filePaths.push_back("../../data/lena.bmp");
+        filePaths.push_back("../../data/ape.png");
+//        filePaths.push_back("../../data/snail.jpg");
+    }
+
+
+    try // <---- do not forget to put your code into try.. catch block!
+    {
+            // Load an image from storage
+        for ( auto it = filePaths.begin(); it != filePaths.end(); ++it )
+        {
+            PenguinV_Image::Image image = File_Operation::Load( *it );
+
+            // If the image is empty it means that the image doesn't exist or the file is not readable
+            if ( image.empty() )
+                throw imageException( std::string("Cannot load ") + *it);
+
+            // Convert to gray-scale image if it's not
+            if ( image.colorCount() != PenguinV_Image::GRAY_SCALE )
+                image = Image_Function::ConvertToGrayScale( image );
+
+            // Threshold image with calculated optimal threshold
+            const std::vector< uint32_t > histogram = Image_Function::Histogram( image );
+            const uint8_t thresholdValue = Image_Function::GetThreshold( histogram );
+            image = Image_Function::Threshold( image, thresholdValue );
+
+            // Save result
+            std::string ext = it->substr( it->length() - 3 );
+            File_Operation::Save( "result." + ext, image );
+        }
+    }
+    catch ( const std::exception & ex ) { // uh-oh, something went wrong!
+        std::cout << ex.what() << ". Press any button to continue." << std::endl;
+        std::cin.ignore();
+        return 1;
+    }
+    catch ( ... ) { // uh-oh, something terrible happen!
+        std::cout << "Generic exception raised. Press any button to continue." << std::endl;
+        std::cin.ignore();
+        return 2;
+    }
+
+    std::cout << "Application ended correctly." << std::endl;
+    return 0;
+}
